Расширил каналы полива до 16 + управление вторым насосом для полива

+++ Main/AbstractModule.h
@@ -279,7 +279,7 @@ class ReservationResolver
 typedef struct
 {
   unsigned long WindowsState; // состояние каналов окон, 4 байта = 32 бита = 16 окон)
-  byte WaterChannelsState; // состояние каналов полива, 1 байт, (8 каналов)
+  uint16_t WaterChannelsState; // состояние каналов полива, 2 байта, (16 каналов)
   byte LightChannelsState; // состояние каналов досветки, 1 байт (8 каналов)
   byte PinsState[16]; // состояние пинов, 16 байт, 128 пинов
